ui: phase 5h — extract ChannelFinder mode handler
Move handleChannelFinderMode (~50 lines: forwards keys to the
finder, dispatches Threads activation / channel select / channel
join depending on the result type) to mode_channel_finder.go.

Add mode_handlers_helpers_test.go (test-only file, mirrors the
Phase 3 services_helpers_test.go pattern) with a thin method
shim so existing tests can keep calling
`app.handleChannelFinderMode(msg)`. Production code calls the
free function directly via the modeHandlers map.

24/24 ui packages green, vet clean.
